requirements.txt
flask>=3.0.0
anthropic>=0.40.0
groq>=0.11.0
python-dotenv>=1.0.0
